test: fix unit tests and lint errors
