+++ packages/editor-ui/src/views/SetupWorkflowFromTemplateView/__tests__/setupTemplate.store.test.ts
@@ -1,233 +1,13 @@
 import { useTemplatesStore } from '@/stores/templates.store';
 import { keyFromCredentialTypeAndName } from '@/utils/templates/templateTransforms';
-import type {
-	TemplateCredentialKey,
-	IWorkflowTemplateNodeWithCredentials,
-} from '@/utils/templates/templateTransforms';
-import type { CredentialUsages } from '@/views/SetupWorkflowFromTemplateView/setupTemplate.store';
-import {
-	getAppCredentials,
-	getAppsRequiringCredentials,
-	useSetupTemplateStore,
-	groupNodeCredentialsByKey,
-} from '@/views/SetupWorkflowFromTemplateView/setupTemplate.store';
+import { useSetupTemplateStore } from '@/views/SetupWorkflowFromTemplateView/setupTemplate.store';
 import { setActivePinia } from 'pinia';
 import * as testData from './setupTemplate.store.testData';
 import { createTestingPinia } from '@pinia/testing';
 import { useCredentialsStore } from '@/stores/credentials.store';
-import { newWorkflowTemplateNode } from '@/utils/testData/templateTestData';
 import { useNodeTypesStore } from '@/stores/nodeTypes.store';
 
-const objToMap = <TKey extends string, T>(obj: Record<TKey, T>) => {
-	return new Map(Object.entries(obj)) as Map<TKey, T>;
-};
-
 describe('SetupWorkflowFromTemplateView store', () => {
-	const nodesByName = {
-		Twitter: {
-			name: 'Twitter',
-			type: 'n8n-nodes-base.twitter',
-			position: [720, -220],
-			parameters: {
-				text: '=Hey there, my design is now on a new product ✨\nVisit my {{$json["vendor"]}} shop to get this cool{{$json["title"]}} (and check out more {{$json["product_type"]}}) 🛍️',
-				additionalFields: {},
-			},
-			credentials: {
-				twitterOAuth1Api: 'twitter',
-			},
-			typeVersion: 1,
-		},
-	} satisfies Record<string, IWorkflowTemplateNodeWithCredentials>;
-
-	describe('groupNodeCredentialsByTypeAndName', () => {
-		it('returns an empty array if there are no nodes', () => {
-			expect(groupNodeCredentialsByKey([])).toEqual(new Map());
-		});
-
-		it('returns credentials grouped by type and name', () => {
-			expect(
-				groupNodeCredentialsByKey([
-					{
-						node: nodesByName.Twitter,
-						requiredCredentials: testData.nodeTypeTwitterV1.credentials,
-					},
-				]),
-			).toEqual(
-				objToMap({
-					'twitterOAuth1Api-twitter': {
-						key: 'twitterOAuth1Api-twitter',
-						credentialName: 'twitter',
-						credentialType: 'twitterOAuth1Api',
-						nodeTypeName: 'n8n-nodes-base.twitter',
-						usedBy: [nodesByName.Twitter],
-					},
-				}),
-			);
-		});
-
-		it('returns credentials grouped when the credential names are the same', () => {
-			const [node1, node2] = [
-				newWorkflowTemplateNode({
-					type: 'n8n-nodes-base.twitter',
-					credentials: {
-						twitterOAuth1Api: 'credential',
-					},
-				}) as IWorkflowTemplateNodeWithCredentials,
-				newWorkflowTemplateNode({
-					type: 'n8n-nodes-base.telegram',
-					credentials: {
-						telegramApi: 'credential',
-					},
-				}) as IWorkflowTemplateNodeWithCredentials,
-			];
-
-			expect(
-				groupNodeCredentialsByKey([
-					{
-						node: node1,
-						requiredCredentials: testData.nodeTypeTwitterV1.credentials,
-					},
-					{
-						node: node2,
-						requiredCredentials: testData.nodeTypeTelegramV1.credentials,
-					},
-				]),
-			).toEqual(
-				objToMap({
-					'twitterOAuth1Api-credential': {
-						key: 'twitterOAuth1Api-credential',
-						credentialName: 'credential',
-						credentialType: 'twitterOAuth1Api',
-						nodeTypeName: 'n8n-nodes-base.twitter',
-						usedBy: [node1],
-					},
-					'telegramApi-credential': {
-						key: 'telegramApi-credential',
-						credentialName: 'credential',
-						credentialType: 'telegramApi',
-						nodeTypeName: 'n8n-nodes-base.telegram',
-						usedBy: [node2],
-					},
-				}),
-			);
-		});
-	});
-
-	describe('getAppsRequiringCredentials', () => {
-		it('returns an empty array if there are no nodes', () => {
-			const appNameByNodeTypeName = () => 'Twitter';
-			expect(getAppsRequiringCredentials(new Map(), appNameByNodeTypeName)).toEqual([]);
-		});
-
-		it('returns an array of apps requiring credentials', () => {
-			const credentialUsages = objToMap<TemplateCredentialKey, CredentialUsages>({
-				[keyFromCredentialTypeAndName('twitterOAuth1Api', 'twitter')]: {
-					key: keyFromCredentialTypeAndName('twitterOAuth1Api', 'twitter'),
-					credentialName: 'twitter',
-					credentialType: 'twitterOAuth1Api',
-					nodeTypeName: 'n8n-nodes-base.twitter',
-					usedBy: [nodesByName.Twitter],
-				},
-			});
-
-			const appNameByNodeTypeName = () => 'Twitter';
-
-			expect(getAppsRequiringCredentials(credentialUsages, appNameByNodeTypeName)).toEqual([
-				{
-					appName: 'Twitter',
-					count: 1,
-				},
-			]);
-		});
-	});
-
-	describe('getAppCredentials', () => {
-		it('returns an empty array if there are no nodes', () => {
-			const appNameByNodeTypeName = () => 'Twitter';
-			expect(getAppCredentials([], appNameByNodeTypeName)).toEqual([]);
-		});
-
-		it('returns an array of apps requiring credentials', () => {
-			const credentialUsages: CredentialUsages[] = [
-				{
-					key: keyFromCredentialTypeAndName('twitterOAuth1Api', 'twitter'),
-					credentialName: 'twitter',
-					credentialType: 'twitterOAuth1Api',
-					nodeTypeName: 'n8n-nodes-base.twitter',
-					usedBy: [nodesByName.Twitter],
-				},
-			];
-
-			const appNameByNodeTypeName = () => 'Twitter';
-
-			expect(getAppCredentials(credentialUsages, appNameByNodeTypeName)).toEqual([
-				{
-					appName: 'Twitter',
-					credentials: [
-						{
-							key: 'twitterOAuth1Api-twitter',
-							credentialName: 'twitter',
-							credentialType: 'twitterOAuth1Api',
-							nodeTypeName: 'n8n-nodes-base.twitter',
-							usedBy: [nodesByName.Twitter],
-						},
-					],
-				},
-			]);
-		});
-	});
-
-	describe('credentialOverrides', () => {
-		beforeEach(() => {
-			setActivePinia(
-				createTestingPinia({
-					stubActions: false,
-				}),
-			);
-
-			const credentialsStore = useCredentialsStore();
-			credentialsStore.setCredentialTypes([testData.credentialTypeTelegram]);
-			const templatesStore = useTemplatesStore();
-			templatesStore.addWorkflows([testData.fullShopifyTelegramTwitterTemplate]);
-			const nodeTypesStore = useNodeTypesStore();
-			nodeTypesStore.setNodeTypes([
-				testData.nodeTypeTelegramV1,
-				testData.nodeTypeTwitterV1,
-				testData.nodeTypeShopifyTriggerV1,
-			]);
-			const setupTemplateStore = useSetupTemplateStore();
-			setupTemplateStore.setTemplateId(testData.fullShopifyTelegramTwitterTemplate.id.toString());
-		});
-
-		it('should return an empty object if there are no credential overrides', () => {
-			// Setup
-			const setupTemplateStore = useSetupTemplateStore();
-
-			expect(setupTemplateStore.credentialUsages.length).toBe(3);
-			expect(setupTemplateStore.credentialOverrides).toEqual({});
-		});
-
-		it('should return overrides for one node', () => {
-			// Setup
-			const credentialsStore = useCredentialsStore();
-			credentialsStore.setCredentials([testData.credentialsTelegram1]);
-
-			const setupTemplateStore = useSetupTemplateStore();
-			setupTemplateStore.setSelectedCredentialId(
-				keyFromCredentialTypeAndName('twitterOAuth1Api', 'twitter'),
-				testData.credentialsTelegram1.id,
-			);
-
-			expect(setupTemplateStore.credentialUsages.length).toBe(3);
-			expect(setupTemplateStore.credentialOverrides).toEqual({
-				'twitterOAuth1Api-twitter': {
-					id: testData.credentialsTelegram1.id,
-					name: testData.credentialsTelegram1.name,
-				},
-			});
-		});
-	});
-
 	describe('setInitialCredentialsSelection', () => {
 		beforeEach(() => {
 			setActivePinia(
